+++ autoelective/__init__.py
@@ -3,6 +3,6 @@
 # filename: __init__.py
 # modified: 2019-09-13
 
-__version__ = "4.0.1"
-__date__    = "2020.05.30"
+__version__ = "5.0.1"
+__date__    = "2020.09.25"
 __author__  = "Rabbit"