+++ autoelective/const.py
@@ -11,9 +11,10 @@ CAPTCHA_CACHE_DIR       = absp("../cache/captcha/")
 LOG_DIR                 = absp("../log/")
 ERROR_LOG_DIR           = absp("../log/error")
 REQUEST_LOG_DIR         = absp("../log/request/")
+WEB_LOG_DIR             = absp("../log/web/")
 
 CNN_MODEL_FILE          = absp("./captcha/model/cnn.pt.gz")
-USER_AGENTS_TXT         = absp("../user_agents.txt")
+USER_AGENTS_TXT_GZ      = absp("../user_agents.txt.gz")
 USER_AGENTS_USER_TXT    = absp("../user_agents.user.txt")
 DEFAULT_CONFIG_INI      = absp("../config.ini")
 
@@ -22,14 +23,12 @@ mkdir(CAPTCHA_CACHE_DIR)
 mkdir(LOG_DIR)
 mkdir(ERROR_LOG_DIR)
 mkdir(REQUEST_LOG_DIR)
+mkdir(WEB_LOG_DIR)
 
-
-# 警惕直接复制的 User-Agent 中可能存在的省略号（例如源自 Firefox 开发者工具），它可能会引发如下错误：
-#   File "/usr/lib/python3.6/http/client.py", line 1212, in putheader
-#     values[i] = one_value.encode('latin-1')
-# UnicodeEncodeError: 'latin-1' codec can't encode character '\u2026' in position 30: ordinal not in range(256)
-USER_AGENT_LIST = read_list(USER_AGENTS_USER_TXT) if os.path.exists(USER_AGENTS_USER_TXT) \
-                    else read_list(USER_AGENTS_TXT)
+if os.path.exists(USER_AGENTS_USER_TXT):
+    USER_AGENT_LIST = read_list(USER_AGENTS_USER_TXT)
+else:
+    USER_AGENT_LIST = read_list(USER_AGENTS_TXT_GZ)
 
 
 class IAAAURL(object):